+++ src/ems.h
@@ -94,7 +94,7 @@ typedef struct {
     bool             emsTxCapable;     // able to send via Tx
     bool             emsTxDisabled;    // true to prevent all Tx
     uint8_t          txRetryCount;     // # times the last Tx was re-sent
-    uint8_t          emsTxDelay;       // handles Tx logic
+    uint8_t          emsTxMode;        // handles Tx logic
     bool             emsReverse;       // if true, poll logic is reversed
 } _EMS_Sys_Status;
 
@@ -302,7 +302,7 @@ void ems_setFlowTemp(uint8_t temperature);
 void ems_setWarmWaterActivated(bool activated);
 void ems_setWarmTapWaterActivated(bool activated);
 void ems_setPoll(bool b);
-void ems_setTxDelay(uint8_t delay);
+void ems_setTxMode(uint8_t mode);
 void ems_setLogging(_EMS_SYS_LOGGING loglevel);
 void ems_setEmsRefreshed(bool b);
 void ems_setWarmWaterModeComfort(uint8_t comfort);
@@ -317,7 +317,7 @@ bool             ems_getPoll();
 bool             ems_getTxEnabled();
 bool             ems_getThermostatEnabled();
 bool             ems_getBoilerEnabled();
-uint8_t          ems_getTxDelay();
+uint8_t          ems_getTxMode();
 bool             ems_getBusConnected();
 _EMS_SYS_LOGGING ems_getLogging();
 bool             ems_getEmsRefreshed();